ci: Fix the condition to run the nox-all and nox-cross-arch-all jobs
If the underlying matrix jobs (nox and nox-cross-arch) fail, the -all
jobs are skipped, but this is not what we want when running in the merge
queue, as we want the merge queue checks to too fail in this case.

So we need to propagate the failures of the matrix jobs to the -all
jobs, they should only be skipped if the matrix jobs were skipped too.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>